Simplified interfaces, improved docs
