+++ samples/04.convex-react-nextjs/app/tasks-state.ts
@@ -1,44 +1,17 @@
 "use client";
 
-import {
-    ConvexQueryNode,
-    createRetreeConvexMutation,
-    RetreeConvexMutation,
-} from "@retreejs/convex";
-import { ReactiveNode } from "@retreejs/core";
+import { ConvexNode, ConvexQueryNode } from "@retreejs/convex";
 import { ConvexClient } from "convex/browser";
 import { api } from "../convex/_generated/api";
 import { Id } from "../convex/_generated/dataModel";
 
-export class TasksState extends ReactiveNode {
+export class TasksState extends ConvexNode {
     public readonly tasks: ConvexQueryNode<typeof api.tasks.get>;
-    private readonly toggleCompletedMutation: RetreeConvexMutation<
-        typeof api.tasks.toggleCompleted
-    >;
-    private readonly closeClient: () => Promise<unknown>;
 
-    constructor(convexUrl: string | undefined) {
-        super();
-        if (convexUrl === undefined) {
-            throw new Error(
-                "TasksState: NEXT_PUBLIC_CONVEX_URL must be defined."
-            );
-        }
-        if (convexUrl.length === 0) {
-            throw new Error(
-                "TasksState: NEXT_PUBLIC_CONVEX_URL cannot be an empty string."
-            );
-        }
-
+    constructor(convexUrl: string) {
         const client = new ConvexClient(convexUrl);
-        this.tasks = new ConvexQueryNode(client, api.tasks.get, {
-            args: {},
-        });
-        this.toggleCompletedMutation = createRetreeConvexMutation(
-            client,
-            api.tasks.toggleCompleted
-        );
-        this.closeClient = () => client.close();
+        super(client);
+        this.tasks = this.query(api.tasks.get);
     }
 
     get dependencies() {
@@ -47,21 +20,23 @@ export class TasksState extends ReactiveNode {
 
     public dispose(): void {
         this.tasks.dispose();
-        void this.closeClient();
+        void this.client.close();
     }
 
     public toggleCompleted(taskId: Id<"tasks">): Promise<null> {
-        return this.toggleCompletedMutation(
+        const toggleCompletedMutation = this.mutation(
+            api.tasks.toggleCompleted
+        );
+        return toggleCompletedMutation(
             { taskId },
             {
-                onOptimistic: (mutation) => {
-                    void this.tasks.applyOptimisticMutation(mutation, {
+                withOptimisticUpdate: (ctx) => {
+                    this.tasks.optimisticUpdate(ctx, {
                         apply(tasks) {
                             const task = tasks.find(
                                 (candidateTask) => candidateTask._id === taskId
                             );
                             if (!task) return;
-
                             task.isCompleted = !task.isCompleted;
                         },
                     });
